Move OrbitalObjectType to its own file
