+++ src/ts/planets/gasPlanet/gasPlanetModel.ts
@@ -23,7 +23,7 @@ import { CelestialBodyModel } from "../../architecture/celestialBody";
 import { newSeededRingsModel } from "../../rings/ringsModel";
 import { GenerationSteps } from "../../utils/generationSteps";
 import { getRngFromSeed } from "../../utils/getRngFromSeed";
-import { OrbitalObjectType } from "../../architecture/orbitalObject";
+import { OrbitalObjectType } from "../../architecture/orbitalObjectType";
 import { PlanetModel } from "../../architecture/planet";
 import { Tools } from "@babylonjs/core/Misc/tools";
