+++ src/ts/starSystem/starSystemController.ts
@@ -27,8 +27,9 @@ import { Star } from "../stellarObjects/star/star";
 import { BlackHole } from "../stellarObjects/blackHole/blackHole";
 import { NeutronStar } from "../stellarObjects/neutronStar/neutronStar";
 import { ChunkForge } from "../planets/telluricPlanet/terrain/chunks/chunkForge";
-import { OrbitalObject, OrbitalObjectType } from "../architecture/orbitalObject";
+import { OrbitalObject } from "../architecture/orbitalObject";
 import { OrbitalObjectUtils } from "../architecture/orbitalObjectUtils";
+import { OrbitalObjectType } from "../architecture/orbitalObjectType";
 import { CelestialBody } from "../architecture/celestialBody";
 import { StellarObject } from "../architecture/stellarObject";
 import { PlanetaryMassObject } from "../architecture/planetaryMassObject";
